begin porting to PHP
This is not even remotely usable.  This is mostly a straight port with a few minor php-specific optimizations.  About 75% of the file is converted.  After each file is converted to PHP, they will be split apart into separate files for each class and put in the src/ directory.  After that, the tests will be ported to PHPUnit and finally it will be ready to roll.